refactor: scaffold
